Move arrow merge to test_merge_disposition
