+++ tests/load/pipeline/test_merge_disposition.py
@@ -20,8 +20,10 @@ from dlt.common.schema.exceptions import (
 from dlt.common.schema.typing import TLoaderMergeStrategy
 from dlt.common.typing import StrAny
 from dlt.common.utils import digest128
-from dlt.common.destination import AnyDestination
+from dlt.common.destination import AnyDestination, DestinationCapabilitiesContext
 from dlt.common.destination.exceptions import DestinationCapabilitiesException
+from dlt.common.libs.pyarrow import row_tuples_to_arrow
+
 from dlt.extract import DltResource
 from dlt.sources.helpers.transform import skip_first, take_first
 from dlt.pipeline.exceptions import PipelineStepFailed
@@ -1521,3 +1523,82 @@ def test_merge_key_null_values(destination_config: DestinationTestConfiguration)
 
     assert isinstance(ex.__context__, NormalizeJobFailed)
     assert isinstance(ex.__context__.__context__, CannotCoerceNullException)
+
+
+@pytest.mark.essential
+@pytest.mark.parametrize(
+    "destination_config",
+    destinations_configs(
+        default_sql_configs=True,
+        local_filesystem_configs=True,
+        table_format_filesystem_configs=True,
+        supports_merge=True,
+        bucket_subset=(FILE_BUCKET),
+    ),
+    ids=lambda x: x.name,
+)
+@pytest.mark.parametrize("merge_strategy", ("delete-insert", "upsert"))
+def test_merge_arrow(
+    destination_config: DestinationTestConfiguration,
+    merge_strategy: TLoaderMergeStrategy,
+) -> None:
+    pipeline = destination_config.setup_pipeline("clickhouse_arrow", dev_mode=True)
+
+    skip_if_not_supported(merge_strategy, pipeline.destination)
+
+    @dlt.resource(
+        write_disposition={"disposition": "merge", "strategy": merge_strategy},
+        primary_key="id",
+    )
+    def arrow_items(rows, schema_columns, timezone="UTC"):
+        yield row_tuples_to_arrow(
+            rows,
+            DestinationCapabilitiesContext.generic_capabilities(),
+            columns=schema_columns,
+            tz=timezone,
+        )
+
+    schema_columns = {
+        "id": {"name": "id", "nullable": False, "data_type": "bigint"},
+        "name": {"name": "name", "nullable": True, "data_type": "text"},
+    }
+    test_rows = [(1, "foo"), (2, "bar")]
+
+    load_info = pipeline.run(
+        arrow_items(test_rows, schema_columns),
+    )
+    assert_load_info(load_info)
+    table_counts = load_table_counts(pipeline, "arrow_items")
+    assert table_counts["arrow_items"] == 2
+
+    tables = load_tables_to_dicts(pipeline, "arrow_items")
+
+    assert_records_as_set(
+        tables["arrow_items"],
+        [
+            {"id": 1, "name": "foo"},
+            {"id": 2, "name": "bar"},
+        ],
+    )
+
+    # Update the records
+    test_rows = [(1, "foo"), (2, "updated bar")]
+
+    load_info = pipeline.run(
+        arrow_items(test_rows, schema_columns),
+    )
+
+    assert_load_info(load_info)
+
+    table_counts = load_table_counts(pipeline, "arrow_items")
+    assert table_counts["arrow_items"] == 2
+
+    tables = load_tables_to_dicts(pipeline, "arrow_items")
+
+    assert_records_as_set(
+        tables["arrow_items"],
+        [
+            {"id": 1, "name": "foo"},
+            {"id": 2, "name": "updated bar"},
+        ],
+    )
